requirements.txt
-i https://pypi.org/simple
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
configparser==3.7.3
idna-ssl==1.1.0 ; python_version < '3.7'
idna==2.8
multidict==4.5.2
rpi.gpio==0.6.5
telepot==12.7
typing-extensions==3.7.2 ; python_version < '3.7'
urllib3==1.24.1
yarl==1.3.0
